Stream docs, int64 conversion.

+++ include/frames/stream.h
@@ -5,29 +5,44 @@
 #define FRAMES_STREAM
 
 #include "frames/noncopyable.h"
+#include "frames/types.h"
 
 #include <string>
 #include <cstdio>
 #include <vector>
 
 namespace Frames {
+  /// Interface for reading serialized data from a generic source.
+  /** Frames Streams do not currently support any sort of asynchronous reading; this is planned for the future. */
   class Stream : detail::Noncopyable {
   public:
     Stream() { };
     virtual ~Stream() { };
 
-    virtual int Read(unsigned char *target, int bytes) = 0;  // returns number of bytes, 0 indicates EOF
-    virtual bool Seek(int offset) = 0;  // true if successful
-    virtual bool Seekable() const = 0; // true if Seek is supported
+    /// Reads a number of bytes.
+    /** Returns the number of bytes successfully read. If the return value is not equal to the bytes parameter then it indicates either EOF or an error condition. */
+    virtual int64_t Read(unsigned char *target, int64_t bytes) = 0;
+
+    /// Seeks the read pointer to a given location in the file.
+    /** Returns true if the read pointer was successfully moved. Failure leaves the read pointer in an undefined position. */
+    virtual bool Seek(int64_t offset) = 0;
+
+    /// Indicates whether Seek is valid on this stream.
+    /** A "false" return value means that all calls to Seek will fail. */
+    virtual bool Seekable() const = 0;
   };
 
+  /// Implementation of Stream reading from standard disk files.
+  /** See Stream for detailed function documentation. */
   class StreamFile : public Stream {
   public:
+    /// Creates a StreamFile referring to a given file.
+    /** Returns NULL if the file does not exist or cannot be read. */
     static StreamFile *Create(const std::string &fname);
     ~StreamFile();
 
-    virtual int Read(unsigned char *target, int bytes);
-    virtual bool Seek(int offset);
+    virtual int64_t Read(unsigned char *target, int64_t bytes);
+    virtual bool Seek(int64_t offset);
     virtual bool Seekable() const;
 
   private:
@@ -36,20 +51,24 @@ namespace Frames {
     std::FILE *m_file;
   };
   
+  /// Implementation of Stream reading from in-memory data.
+  /** See Stream for detailed function documentation. */
   class StreamBuffer : public Stream {
   public:
+    /// Creates a StreamBuffer referring to given data.
+    /** For simplicity, this data is copied inside the StreamBuffer. */
     static StreamBuffer *Create(const std::vector<unsigned char> &data);
     ~StreamBuffer();
 
-    virtual int Read(unsigned char *target, int bytes);
-    virtual bool Seek(int offset);
+    virtual int64_t Read(unsigned char *target, int64_t bytes);
+    virtual bool Seek(int64_t offset);
     virtual bool Seekable() const;
 
   private:
     StreamBuffer(const std::vector<unsigned char> &data);
 
     std::vector<unsigned char> m_data;
-    int m_index;
+    int64_t m_index;
   };
 }
